Refactored Login Register
