+++ acm_website/backend/.env
@@ -1,0 +1,2 @@
+SESSION_SECRET_KEY=93cfdb1e9338c5eaa1a9a77d915763558fb0d1832733246ac2ffaf205f7759297131da2d847681f12f4e6e446b83a11fd04f33ee9ca1342a54a042593d8e2b4e
+DATABASE_URL=mongodb+srv://pollBoss:p0llM%40steR@cluster0.j4fxq.mongodb.net/acm_website?retryWrites=true&w=majority
