+++ acm_website/backend/middleware/auth.js
@@ -1,26 +1,35 @@
-const config = require('config')
+//const config = require('config')
 
 const jwt = require('jsonwebtoken')
 const { models } = require('mongoose')
 
+// function auth(req,res,next) {
+//     const token = req.header('x-auth-token')
+
+//     // check for token
+//     if(!token) res.status(401).json({ msg:'No token:Authorization denied '})
+
+//     try{
+
+//         // Verify token
+//         const decoded = jwt.verify(token,config.get('jwtSecret'))
+
+//         // Add user from payload
+//         req.user = decoded
+//         next();
+//     }catch(err) {
+//         res.status(400).json({ msg:'token is not valid' })
+//     }
+
+// }
+
 function auth(req,res,next) {
-    const token = req.header('x-auth-token')
-
-    // check for token
-    if(!token) res.status(401).json({ msg:'No token:Authorization denied '})
-
-    try{
-
-        // Verify token
-        const decoded = jwt.verify(token,config.get('jwtSecret'))
-
-        // Add user from payload
-        req.user = decoded
+    if(req.isAuthenticated()){
         next();
-    }catch(err) {
-        res.status(400).json({ msg:'token is not valid' })
     }
-
+    else {
+        res.status(401).json({ msg:'Not authenticated' })
+    }
 }
 
 module.exports = auth
